Update the collection api endpoint by adding a new field products_count

+++ store/views.py
@@ -1,5 +1,6 @@
-from os import stat
+from django.db.models.aggregates import Count
 from django.shortcuts import render, get_object_or_404
+
 from django.http import HttpResponse
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
@@ -52,7 +53,8 @@ def product_detail(request, id):
 @api_view(['GET', 'POST'])
 def collection_list(request):
     if request.method == 'GET':
-        queryset = Collection.objects.all()
+        queryset = Collection.objects.annotate(
+            products_count=Count('products')).all()
         serializer = CollectionSerializer(queryset, many=True)
         return Response(serializer.data)
     elif request.method == 'POST':
@@ -64,7 +66,9 @@ def collection_list(request):
 
 @api_view(['GET', 'PUT', 'DELETE'])
 def collection_detail(request, pk):
-    collection = get_object_or_404(Collection, pk=pk)
+    collection = get_object_or_404(
+        Collection.objects.annotate(
+            products_count=Count('products')), pk=pk)
     if request.method == 'GET':
         serializer = CollectionSerializer(collection)
         return Response(serializer.data)
@@ -74,5 +78,9 @@ def collection_detail(request, pk):
         serializer.save()
         return Response(serializer.data)
     elif request.method == 'DELETE':
+        if collection.products.count() > 0:
+            return Response(
+                {'error': 'Collection cannot be delete as it includes one or more products'},
+                status=status.HTTP_405_METHOD_NOT_ALLOWED)
         collection.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
